personal_app/requirements.txt
Django>=5.0,<6.0
torch>=2.1
sentence-transformers>=3.0
PyYAML>=6.0
